+++ Ico/DefaultsManager.swift
@@ -14,7 +14,7 @@ class DefaultsManager {
 
     var polyhedronName: String {
         get {
-            return defaults.string(forKey: "polyhedron_name") ?? Polyhedra.defaultName
+            return defaults.string(forKey: "polyhedron_name") ?? PolyhedraRegistry.defaultName
         }
         set(value) {
             defaults.setValue(value, forKey: "polyhedron_name")
